+++ tests/deprecated_api/test_remove_1-6.py
@@ -17,7 +17,6 @@ from unittest.mock import call, Mock
 import pytest
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.plugins.training_type import DDPPlugin
 from pytorch_lightning.utilities.distributed import rank_zero_deprecation, rank_zero_warn
 from pytorch_lightning.utilities.model_helpers import is_overridden
@@ -71,14 +70,6 @@ def test_v1_6_0_is_overridden_model():
         assert is_overridden("validation_step", model=model)
     with pytest.deprecated_call(match="and will be removed in v1.6"):
         assert not is_overridden("foo", model=model)
-
-
-def test_v1_6_0_early_stopping_monitor(tmpdir):
-    with pytest.deprecated_call(
-        match=r"The `EarlyStopping\(monitor\)` argument will be required starting in v1.6."
-        " For backward compatibility, setting this to `early_stop_on`."
-    ):
-        EarlyStopping()
 
 
 def test_v1_6_0_train_loop(tmpdir):